backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
sqlalchemy[asyncio]==2.0.18
asyncpg==0.27.0
alembic==1.11.1
pydantic-settings==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.0